🚚 rename: MyTeamList를 index로 변경 #50
